move getRowMetadata & setRowMetadata to DataMode.js
