+++ src/dataModels/HeaderSubgrid.js
@@ -1,5 +1,7 @@
 'use strict';
 
+var DataModel = require('./DataModel');
+
 /**
  * @implements dataModelAPI
  * @param {Hypergrid} grid
@@ -7,22 +9,19 @@
  * @param {string} [options.name]
  * @constructor
  */
-function HeaderSubgrid(grid, options) {
-    this.grid = grid;
-    this.behavior = grid.behavior;
+var HeaderSubgrid = DataModel.extend('HeaderSubgrid', {
+    initialize: function(grid, options) {
+        this.behavior = grid.behavior;
 
-    /**
-     * @type {dataRowObject}
-     */
-    this.dataRow = {}; // for meta data (__HEIGHT)
+        /**
+         * @type {dataRowObject}
+         */
+        this.dataRow = {}; // for meta data (__HEIGHT)
 
-    if (options && options.name) {
-        this.name = options.name;
-    }
-}
-
-HeaderSubgrid.prototype = {
-    constructor: HeaderSubgrid.prototype.constructor,
+        if (options && options.name) {
+            this.name = options.name;
+        }
+    },
 
     type: 'header',
 
@@ -45,15 +44,7 @@ HeaderSubgrid.prototype = {
 
     getRow: function(y) {
         return this.dataRow;
-    },
-
-    getRowMetadata: function(y, metadata) {
-        return this.metadata || (this.metadata = metadata);
-    },
-
-    setRowMetadata: function(y, metadata) {
-        return (this.metadata = metadata);
     }
-};
+});
 
 module.exports = HeaderSubgrid;
